@@ -14,28 +14,23 @@ classifiers =
     Development Status :: 5 - Production/Stable
     Environment :: Web Environment
     Framework :: Django
-    Framework :: Django :: 2.2
-    Framework :: Django :: 3.0
-    Framework :: Django :: 3.1
     Framework :: Django :: 3.2
-    Framework :: Django :: 4.0
     Framework :: Django :: 4.1
+    Framework :: Django :: 4.2
     Intended Audience :: Developers
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
 
 [options]
 packages = find:
 install_requires =
     django-js-asset
-python_requires = >=3.6
+python_requires = >=3.9
 include_package_data = True
 zip_safe = False
 
@@ -48,17 +43,6 @@ tests =
 exclude =
     tests
     tests.*
-
-[flake8]
-exclude = venv,.tox,build,docs
-ignore = E501,F841,W503
-max-line-length = 88
-# max-complexity = 10
-
-[isort]
-profile = black
-combine_as_imports = True
-lines_after_imports = 2
 
 [coverage:run]
 branch = True
