@@ -4,9 +4,9 @@ from myapp.models import Category, Person
 from mptt.admin import DraggableMPTTAdmin, MPTTModelAdmin
 
 
+@admin.register(Category)
 class CategoryAdmin(MPTTModelAdmin):
     pass
 
 
-admin.site.register(Category, CategoryAdmin)
 admin.site.register(Person, DraggableMPTTAdmin)
